Added option to disable registering cameras (#33)
Arlo now released HomeKit support for Arlo Pro and Pro 2. But the support only provides access to the cameras and motion detection, but not the HomeKit Security System feature. This HomeBridge plugin is still very useful but users might not want to include the cameras, as they would be available twice then.

To support this scenario this commit adds a new optional configuration parameter ‘include_cameras” (default is set to true) which will exclude any camera from being registered via HomeBridge. Another option would have been to add a whitelist or blacklist feature for users, but I didn’t want this to be too complex without knowing the need for this.