Camera Streaming (#43)
* Adapt homebridge-camera-ffmpeg source

* Remove standalone service function

* Add some debugging calls

* Use self for callback

* Add some debugging for ffmpeg

* Add required spawn module

* Use name instead of ID for device name

* Add api check

* Split ArloCameraSource into separate file to handle license better

* Manage options at constructor rather than at stream configuration

* Allow skipping of video transcoding if requested video size matches source

* Add ip and debug modules to package

* Add documentation to README about streaming

* Switch to libopus for audio transcoding due to substantially better performance, also enable comfort noise on stream

* Allow for additional audio commands if user desires

* Some additional documentation and note that libopus is now default audio codec

* Update `ffmpeg` (#41)

* Update `ffmpeg`

Add `ffmpeg-for-homebridge` - static `ffmpeg` binaries for Homebridge with support for audio (`libfdk-aac`) and hardware decoding (`h264_omx`).

* Create config.schema.json

Co-authored-by: Donavan Becker <9875439+donavanbecker@users.noreply.github.com>

* Added option to disable registering cameras (#33)

Arlo now released HomeKit support for Arlo Pro and Pro 2. But the support only provides access to the cameras and motion detection, but not the HomeKit Security System feature. This HomeBridge plugin is still very useful but users might not want to include the cameras, as they would be available twice then.

To support this scenario this commit adds a new optional configuration parameter ‘include_cameras” (default is set to true) which will exclude any camera from being registered via HomeBridge. Another option would have been to add a whitelist or blacklist feature for users, but I didn’t want this to be too complex without knowing the need for this.

Co-authored-by: Charles Powell <cbpowell@gmail.com>
Co-authored-by: David Parry <david@smugmug.com>
Co-authored-by: Kristian Matthews <kristian.matthews@me.com>
Co-authored-by: Andreas Linde <TheRealKerni@users.noreply.github.com>